feat: Balance 충전 및 조회 기능 추가

+++ src/main/java/com/practice/commerce/api/dto/request/BalanceRequest.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @Schema(description = "잔액 충전 요청을 위한 데이터 전송 객체")
 public class BalanceRequest {
 	@Schema(description = "충전 또는 소진할 금액", example = "100.00", required = true)
-	private Double amount;
+	private long amount;
 }
